Cast values out of the DbDataRecord instead of parsing .ToString to avoid typeing mismatches

+++ MystiickWeb.Shared/Models/CameraSettings.cs
@@ -4,7 +4,7 @@ public class CameraSettings
 {
     public string Model { get; init; } = string.Empty;
     public bool Flash { get; init; }
-    public uint ISO { get; init; }
+    public short ISO { get; init; }
     public string ShutterSpeed { get; init; } = string.Empty;
     public string Aperature { get; init; } = string.Empty;
     public string FocalLength { get; init; } = string.Empty;
